Refactor Table.update tests
